feat: remove the last applied annotation on upgrade
